moved apps to /src
