+++ src/RayleighBenardConvection/Convection_Dash.jl
@@ -1,35 +1,39 @@
-module RTITools
-
+module ConvectionTools
 using Dash, DashBootstrapComponents
 using PlotlyJS
 using LaMEM
 using UUIDs
 using Interpolations
 using GeophysicalModelGenerator
+using DelimitedFiles
 using HTTP
 
-export RayleighTaylor
+export Convection 
 
-pkg_dir = Base.pkgdir(RTITools)
+pkg_dir = Base.pkgdir(ConvectionTools)
 
 include(joinpath(pkg_dir,"src/dash_tools.jl"))
-include(joinpath(pkg_dir,"RayleighTaylorInstability/dash_functions_RTI.jl"))
+include(joinpath(pkg_dir,"src/RayleighBenardConvection/dash_functions_convection.jl"))
  
-"""
-    RayleighTaylor(; host=HTTP.Sockets.localhost, port=8050)
 
-This starts a RayleighTaylor instability GUI
 """
-function RayleighTaylor(; host = HTTP.Sockets.localhost, port=8050)
-    pkg_dir = Base.pkgdir(RTITools)
+    Convection(; host=HTTP.Sockets.localhost, port=8050)
+
+This starts a convection GUI
+"""
+function Convection(; host=HTTP.Sockets.localhost, port=8050)
+    pkg_dir = Base.pkgdir(ConvectionTools)
+    cmaps = read_colormaps(dir_colormaps=joinpath(pkg_dir,"src/assets/colormaps/"))
     
     GUI_version = "0.1.1"
-    cmaps = read_colormaps(dir_colormaps=joinpath(pkg_dir,"src/assets/colormaps/"))
-
-    title_app = "Rayleigh Taylor Instability"
-    ParamFile = "RTI.dat"
-    OutFile = "RTI"
     
+    title_app = "Rayleigh-Benard Convection"
+    ParamFile = "Convection.dat"
+    OutFile = "Convection"
+
+    base_dir = pwd()
+    cd(base_dir)
+
     #app = dash(external_stylesheets=[dbc_themes.CYBORG])
     app = dash(external_stylesheets = [dbc_themes.BOOTSTRAP, dbc_icons.BOOTSTRAP], prevent_initial_callbacks=false)
     app.title = title_app
@@ -85,64 +89,53 @@ function RayleighTaylor(; host = HTTP.Sockets.localhost, port=8050)
         Input("button-run", "disabled"),
         Input("button-play", "n_clicks"), 
         State("domain_width", "value"),
-        State("depth", "value"),
+        State("domain_height", "value"),
         State("nel_x", "value"),
         State("nel_z", "value"),
         State("n_timesteps", "value"),
-        State("switch-FreeSurf", "value"),
-        State("switch-Layers", "value"),
+        State("ΔT", "value"),
+        State("γ", "value"),
+        State("cohesion", "value"),
+        State("viscosity", "value"),
         State("last_timestep", "data"),
         State("plot_field", "value"),
-        State("session-id", "data"), 
-        State("viscosity_upper", "value"),
-        State("viscosity_lower", "value"),
-        State("density_upper", "value"),
-        State("density_lower", "value"),
+        State("session-id", "data"),
+        State("switch-FreeSurf","value"),
         prevent_initial_call=true
     ) do n_run, active_run, n_play,
-        domain_width, depth, nel_x, nel_z, n_timesteps,
-        open_top, layers,
-        last_timestep, plot_field, session_id,
-        η_up,η_lo,ρ_up,ρ_lo
+        domain_width, domain_height, nel_x, nel_z, n_timesteps,
+        ΔT, γ, cohesion, viscosity,
+        last_timestep, plot_field, session_id, FreeSurface
         
-
-        # print(layers)
-        # print(open_top)
-
         trigger = get_trigger()
         disable_interval = true
+        user_dir = simulation_directory(session_id, clean=false)
         if trigger == "button-run.n_clicks"
             cur_dir = pwd()
-            base_dir = joinpath(pkgdir(RTITools),"RayleighTaylorInstability")
-            
-            
-            η_up = 10.0^η_up
-            η_lo = 10.0^η_lo
-            Hi_value = depth
-            W    = domain_width
-            open_top_bound = active_switch(open_top)
+            base_dir = joinpath(pkgdir(ConvectionTools),"src","RayleighBenardConvection")
 
-            # print(open_top)
-            addlayers = active_switch(layers)
-            
-            # print(layers)c
+            viscosity = 10.0^viscosity
+            cohesion *= 1.0e6
 
-            args = "-nstep_max $(n_timesteps) -eta[0] $η_up -eta[1] $η_up -eta[2] $η_lo -rho[0] $ρ_up -rho[1] $ρ_up -rho[2] $ρ_lo -open_top_bound $(Int64(open_top_bound)) -nel_x $nel_x -nel_z $nel_z -coord_x $(-W/2),$(W/2)"
+            Δx  = domain_width/nel_x # y-width
+            
+            if FreeSurface === nothing
+                args = "-nstep_max $(n_timesteps) -eta_fk[0] $(viscosity)  -gamma_fk[0] $γ -TRef_fk[0] $(ΔT/2) -ch[0] $(cohesion) -nel_x $nel_x -nel_z $nel_z -coord_x $(-domain_width/2),$(domain_width/2) -coord_z $(-domain_height),0 -coord_y $(-Δx/2),$(Δx/2) -temp_bot $ΔT"  
+            else
+                args = "-nstep_max $(n_timesteps) -eta_fk[0] $(viscosity)  -gamma_fk[0] $γ -TRef_fk[0] $(ΔT/2) -ch[0] $(cohesion) -nel_x $nel_x -coord_x $(-domain_width/2),$(domain_width/2) -coord_y $(-Δx/2),$(Δx/2) -temp_bot $ΔT"
+            end
+
             println("args = ", args)
-            
+
             # We clicked the run button
             user_dir = simulation_directory(session_id, clean=true)
             cd(user_dir)
-
+            
             pfile = joinpath(base_dir,ParamFile) 
-
-            # Create the setup
-            CreateSetup(pfile, addlayers, Hi_value, args=args)
-    
+            CreateSetup(pfile, ΔT, args=args)
             run_lamem(pfile, 1, args, wait=false)
-            cd(cur_dir)        # return to main directory
-
             disable_interval = false
+            cd(cur_dir)        # return to main directory
 
         elseif trigger == "button-run.disabled"
             last_t = parse(Int, last_timestep)
@@ -159,6 +152,7 @@ function RayleighTaylor(; host = HTTP.Sockets.localhost, port=8050)
         return disable_interval
     end
 
+    
     # deactivate the button 
     callback!(app,
         Dash.Output("button-run", "disabled"),
@@ -261,7 +255,7 @@ function RayleighTaylor(; host = HTTP.Sockets.localhost, port=8050)
         trigger == "button-play.n_clicks"
 
             user_dir = simulation_directory(session_id, clean=false)
-            if has_pvd_file(OutFile, user_dir)
+            if  has_pvd_file(OutFile, user_dir)
                 Timestep, _, Time = Read_LaMEM_simulation(OutFile, user_dir)      # all timesteps
                 id = findall(Timestep .== cur_t)[1]
                 if trigger == "button-start.n_clicks" || trigger == "button-play.n_clicks"
@@ -286,16 +280,18 @@ function RayleighTaylor(; host = HTTP.Sockets.localhost, port=8050)
                 else
                     x_con, y_con, data_con = x, y, data
                 end
-
+                
                 # update the plot
                 add_velocity = active_switch(switch_velocity)
-                
+
+
                 fig_cross = create_main_figure(OutFile, cur_t, x, y, data, x_con, y_con, data_con;
                     add_contours=add_contours, contour_field=contour_field,
                     add_velocity=add_velocity,
                     colorscale=color_map_option,
-                    session_id=session_id,
-                    field=plot_field, cmaps=cmaps)
+                    session_id=session_id, 
+                    cmaps=cmaps,
+                    field=plot_field)
 
                 if trigger == "current_timestep.data" || trigger == "update_fig.data" || trigger == "button-play.n_clicks"
                     if cur_t < last_t
@@ -318,7 +314,7 @@ function RayleighTaylor(; host = HTTP.Sockets.localhost, port=8050)
         current_timestep = "$cur_t"
 
         # @show current_timestep
-        println("Timestep ", current_timestep)
+        println("Timestep ", cur_t)
         return label_timestep, label_time, current_timestep, fig_cross, fields_available, fields_available
     end
 
@@ -337,9 +333,12 @@ function RayleighTaylor(; host = HTTP.Sockets.localhost, port=8050)
         end
         return disable_contours
     end
+    
 
     run_server(app, host, port, debug=false)
 
+
+    return app
 end
 
 end
